Many functions added
If endstop topic is not publishing and one motor requires an endstop, the motor torque is set off.
Now we can set a joint proportion between radians/meters, for prismatic joints.
We can set an initial position. This can be usefull if we want always to start in some pose different of 0.